+++ bluetti_mqtt/core/devices/ac500.py
@@ -72,6 +72,7 @@ class AC500(BluettiDevice):
 
         # Battery Data
         self.struct.add_uint_field('pack_num_max', 91)
+        self.struct.add_decimal_field('total_battery_voltage', 92, 1)
         self.struct.add_uint_field('pack_num', 96)
         self.struct.add_decimal_field('pack_voltage', 98, 2)  # Full pack voltage
         self.struct.add_uint_field('pack_battery_percent', 99)